% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{combine}
\alias{combine}
\alias{combine.default}
\alias{combine.data.frame}
\alias{combine.factor}
\alias{combine.formula}
\alias{combine.list}
\alias{combine.logical}
\alias{combine.matrix}
\alias{combine.numeric}
\title{Test Functions}
\usage{
combine(...)

\method{combine}{default}(...)

\method{combine}{data.frame}(...)

\method{combine}{factor}(...)

\method{combine}{formula}(...)

\method{combine}{list}(...)

\method{combine}{logical}(...)

\method{combine}{matrix}(...)

\method{combine}{numeric}(...)
}
\arguments{
\item{...}{arguments passed to other methods.}
}
\description{
Functions to demonstrate R CMD check warning produced by registration of a
formula method when it and its S3 generic both have an ellipsis as the first
argument.
}
